==> part_i_data_exploration_and_statistics/scripts_final/05_plot_wagon_movement_time.ipynb <==
### Imports
import os
import glob

import pandas as pd
import dask.dataframe as dd
import numpy as np

import matplotlib as mpl
import matplotlib.pyplot as plt
import seaborn as sns

plt.style.use(['mystyle'])

# Auswertung der Wagon Moving Time über der Zeit (How often are wagons moved)

## Zeitdaten Plot: tägliche Moving time über jeden Tag/Monat. Für alle Wagon Typen.

# Einlesen der Daten:
df = pd.read_csv(os.path.join("..", "output", "group_01", "wagon_types_movent_data.csv") )
df = df.sort_values(by =["wagon_type", "day"])
df

# add week and month timestamps:
week_days = 7
bins = np.arange(0, max(df.day)+week_days, week_days)
bin_labels = [i+1 for i, _ in enumerate(bins)]
bin_labels.pop()
df['week'] = pd.cut(df['day'], bins = bins, labels = bin_labels)
weeks = bin_labels

month_weeks = 4
bins = np.arange(0, max(df.week)+month_weeks, month_weeks)
bin_labels = [i+1 for i, _ in enumerate(bins)]
bin_labels.pop()
df['month'] = pd.cut(df['week'], bins = bins, labels = bin_labels)
months = bin_labels

df

## mittelung über Monate:
wgTypes = [i for i in range(0,9)]   #wgType=0: all wagons, else wgType = number of wagon type (1-8)

result=[]

for wgType in wgTypes:
    if wgType == 0:
        df_temp = df
    else:
        df_temp = df[df.wagon_type == wgType]

    wgType_weekly_movingTime = []
    wgType_weekly_standingTime = []
    wgType_weekly_parkingTime = []
    for month in months:
        df_week = df_temp[df_temp.month == month]
        wgType_weekly_movingTime.append(df_week.moving_time_mean.mean() )
        wgType_weekly_standingTime.append(df_week.standing_time_mean.mean() )
        wgType_weekly_parkingTime.append(df_week.parking_time_mean.mean() )
    
    result.append([wgType_weekly_movingTime, wgType_weekly_standingTime, wgType_weekly_parkingTime])

#result

### plot: mean weekly moving time over the time (months)
#fig, axs = plt.subplots(3, 1, figsize=(20, 16), facecolor='w', edgecolor='k', constrained_layout=True) #make subplots
#fig.subplots_adjust(hspace = .35, wspace=.1) #height and width space between the subfigures
#axs = axs.ravel()
#
#wgTypes = [i for i in range(0,9)]   # list of wagon types from 1 to 8
#
## plotStyle
#lw = 1.75
#plotTitles = ["monatlich gemittelte tägliche Bewegungszeit", "monatlich gemittelte tägliche Standzeit", "monatlich gemittelte tägliche Parkzeit"]
##linestyles = ["solid", "dashed", "dotted", "dashdot", "densely dashed", "densely dotted", "densely dashdotted", 'loosely dashdotdotted']    #see https://matplotlib.org/stable/gallery/lines_bars_and_markers/linestyles.html
#linestyles = ["-",'-', '--', ':', '-.', (0, (5, 1)), (0, (1, 1)), (0, (3, 1, 1, 1)),  "--"]
#colors = ["r", "tab:blue", "tab:orange", "tab:green", "tab:red", "tab:purple", "tab:brown", "tab:pink", "tab:gray"]
#
#for tile in range(3):   #plot for 3 Subplots moving/standing/parking
#    for wgType in wgTypes: #plot for each wagon:
#        x = months
#        if tile == 0:
#            y = result[wgType][0]
#        elif tile == 1:
#            y = result[wgType][1]
#        elif tile == 2:
#            y = result[wgType][2]
#
#        y = [i/3600 for i in y] #y-achse: Moving/Standing/Parking time in Stunden
#        if wgType == 0:
#            axs[tile].plot(x,y, linewidth=2.5*lw, linestyle=linestyles[wgType], color=colors[wgType], alpha=0.5)
#        else:
#            axs[tile].plot(x,y, linewidth=lw, linestyle=linestyles[wgType], color=colors[wgType])
#
#    #some plot Settings:
#    axs[tile].grid(True)
#    axs[tile].set_xticks(np.arange(1, 13, step=1))
#    axs[tile].set_xlim([1,12])
#    axs[tile].set_ylim([0, axs[tile].get_ylim()[1]*1.2])
#    axs[tile].set_title(plotTitles[tile])
#    axs[tile].legend(["alle Wagons","Typ 1", "Typ 2", "Typ 3", "Typ 4", "Typ 5", "Typ 6", "Typ 7", "Typ 8"], loc='upper center', ncol=9)
#
#plt.setp(axs, ylabel="Zeit in Stunden", xlabel="Monat")
##plt.xscale('linear')
#plt.show()
#
#
#fig.savefig(os.path.join("..","output","plots","group_01","mean_daily_moving_time_over_the_months.png"), \
#    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')
#fig.savefig(os.path.join("..","output","plots","group_01","mean_daily_moving_time_over_the_months.pdf"), \
#    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')



fz = 2*10
fz_lab = 2*8
fz_tick= 2*8
fz_leg = 2*6


## plot: mean weekly moving time over the time (months)
fig, axs = plt.subplots(3, 1, figsize=(14, 10), facecolor='w', edgecolor='k', constrained_layout=True, sharex=True) #make subplots
fig.subplots_adjust(hspace = .25, wspace=.1) #height and width space between the subfigures
axs = axs.ravel()

wgTypes = [i for i in range(0,9)]   # list of wagon types from 1 to 8

# plotStyle
lw = 1.75
plotTitles = ["monatlich gemittelte tägliche Bewegungszeit", "monatlich gemittelte tägliche Standzeit", "monatlich gemittelte tägliche Parkzeit"]
#linestyles = ["solid", "dashed", "dotted", "dashdot", "densely dashed", "densely dotted", "densely dashdotted", 'loosely dashdotdotted']    #see https://matplotlib.org/stable/gallery/lines_bars_and_markers/linestyles.html
linestyles = ["-",'-', '--', ':', '-.', (0, (5, 1)), (0, (1, 1)), (0, (3, 1, 1, 1)),  "--"]
colors = ["r", "tab:blue", "tab:orange", "tab:green", "tab:red", "tab:purple", "tab:brown", "tab:pink", "tab:gray"]

for tile in range(3):   #plot for 3 Subplots moving/standing/parking
    for wgType in wgTypes: #plot for each wagon:
        x = months
        if tile == 0:
            y = result[wgType][0]
        elif tile == 1:
            y = result[wgType][1]
        elif tile == 2:
            y = result[wgType][2]

        y = [i/3600 for i in y] #y-achse: Moving/Standing/Parking time in Stunden
        if wgType == 0:
            axs[tile].plot(x,y, linewidth=2.5*lw, linestyle=linestyles[wgType], color=colors[wgType], alpha=0.5)
        else:
            axs[tile].plot(x,y, linewidth=lw, linestyle=linestyles[wgType], color=colors[wgType])

    #some plot Settings:
    axs[tile].grid(True)
    axs[tile].set_xticks(np.arange(1, 13, step=1))
    axs[tile].tick_params(axis='both', which='major', labelsize=fz_tick)
    axs[tile].tick_params(axis='both', which='minor', labelsize=fz_tick*0.8)

    yticks = [np.arange(0, 4.5, step=1), np.arange(0, 4, step=1), np.arange(0, 17, step=2.5)]
    axs[tile].set_yticks(yticks[tile])

    axs[tile].set_xlim([1,12])
    axs[tile].set_ylim([0, axs[tile].get_ylim()[1]*1.2])
    axs[tile].set_title(plotTitles[tile], fontsize=fz)
    axs[tile].legend(["alle Wagons","Typ 1", "Typ 2", "Typ 3", "Typ 4", "Typ 5", "Typ 6", "Typ 7", "Typ 8"], loc='upper center', ncol=9, fontsize=fz_leg, handletextpad=0.5)
    axs[tile].set_ylabel("Zeit in Stunden", fontsize=fz_lab)

axs[tile].set_xlabel("Monat", fontsize=fz_lab)


#plt.xscale('linear')
plt.show()


fig.savefig(os.path.join("..","output","plots","group_01","mean_daily_moving_time_over_the_months.png"), \
    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')
fig.savefig(os.path.join("..","output","plots","group_01","mean_daily_moving_time_over_the_months.pdf"), \
    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')


# Visualize: plot mean or mean moving/standing/parking time of all wagons over the time (days) -> 3 plots (moving/standing/parking)
fig, axs = plt.subplots(3, 1, figsize=(20, 16), facecolor='w', edgecolor='k') #make subplots
fig.subplots_adjust(hspace = .35, wspace=.1) #height and width space between the subfigures
axs = axs.ravel()

wgTypes = [i for i in range(1,9)]   # list of wagon types from 1 to 8
days = df["day"].unique().tolist()

# plotStyle
lw = 1.75
plotTitles = ["mean daily Moving Time", "mean daily Standing Time", "mean daily Parking Time"]
#linestyles = ["solid", "dashed", "dotted", "dashdot", "densely dashed", "densely dotted", "densely dashdotted", 'loosely dashdotdotted']    #see https://matplotlib.org/stable/gallery/lines_bars_and_markers/linestyles.html
linestyles = ['-', '--', ':', '-.', (0, (5, 1)), (0, (1, 1)), (0, (3, 1, 1, 1)), (0, (3, 10, 1, 10, 1, 10)) ]

for tile in range(3):   #plot for 3 Subplots moving/standing/parking
    for wgType in wgTypes: #plot for each wagon:
        x = df[df["wagon_type"] == wgType]["day"]   #x-achse: Tage von ab Tag 1
        if tile == 0:
            y = df[df["wagon_type"] == wgType]["moving_time_mean"]
            #err = df[df["wagon_type"] == wgType]["moving_time_std"].divide(3600)
        elif tile == 1:
            y = df[df["wagon_type"] == wgType]["standing_time_mean"]
            #err = df[df["wagon_type"] == wgType]["standing_time_std"].divide(3600)
        elif tile == 2:
            y = df[df["wagon_type"] == wgType]["parking_time_mean"]
            #err = df[df["wagon_type"] == wgType]["parking_time_std"].divide(3600)

        y = y.divide(3600) #y-achse: Moving/Standing/Parking time in Stunden
        axs[tile].plot(x,y, linewidth=lw, linestyle=linestyles[wgType-1])
        
        #plot error bars / error regions:
        #axs[tile].fill_between(x, y.subtract(err), y.add(err), alpha=0.075)

    #some plot Settings:
    #axs[tile].set_ylim([0, axs[tile].get_ylim()[1]*1.3])
    axs[tile].set_title(plotTitles[tile])
    axs[tile].legend(["wgType 1", "wgType 2", "wgType 3", "wgType 4", "wgType 5", "wgType 6", "wgType 7", "wgType 8"], loc='upper center', ncol=8)

plt.setp(axs, ylabel="time in hours (per day)", xlabel="day")
plt.xscale('linear')
plt.show()



from numpy import sign, zeros
from scipy.interpolate import interp1d


# plot for/by each wagon type: moving/standing/parking time for each type -> 8 plots
fig, axs = plt.subplots(8, 1, figsize=(18, 40), facecolor='w', edgecolor='k') #make subplots
fig.subplots_adjust(hspace = .30, wspace=.25) #height and width space between the subfigures
axs = axs.ravel()

# plotStyle
lw = 1.75
linestyles = ['-', '--', ':', (0, (5, 10))]

wgTypes = [i for i in range(1,9)]
for tile in range(8):
    wgType = wgTypes[tile]
    x = df[df["wagon_type"] == wgType]["day"]   #x-achse: Tage von ab Tag 1
    y1 = df[df["wagon_type"] == wgType]["moving_time_mean"].divide(3600)
    y2 = df[df["wagon_type"] == wgType]["standing_time_mean"].divide(3600)
    y3 = df[df["wagon_type"] == wgType]["parking_time_mean"].divide(3600)
    axs[tile].plot(x,y1, linewidth=lw, linestyle=linestyles[0])
    axs[tile].plot(x,y2, linewidth=lw, linestyle=linestyles[1])
    axs[tile].plot(x,y3, linewidth=lw, linestyle=linestyles[2])
    
    ## plot envelope for moving time, to mark up seasonal trends:
    # source: https://stackoverflow.com/questions/34235530/how-to-get-high-and-low-envelope-of-a-signal/34245942#34245942
    s = y1.to_numpy()
    q_u = zeros(s.shape)
    q_l = zeros(s.shape)
    #Prepend the first value of (s) to the interpolating values. This forces the model to use the same starting point for both the upper and lower envelope models.
    u_x = [0,]
    u_y = [s[0],]

    l_x = [0,]
    l_y = [s[0],]
    #Detect peaks and troughs and mark their location in u_x,u_y,l_x,l_y respectively.
    for k in range(1,len(s)-1):
        if (sign(s[k]-s[k-1])==1) and (sign(s[k]-s[k+1])==1):
            u_x.append(k)
            u_y.append(s[k])

        if (sign(s[k]-s[k-1])==-1) and ((sign(s[k]-s[k+1]))==-1):
            l_x.append(k)
            l_y.append(s[k])
    #Append the last value of (s) to the interpolating values. This forces the model to use the same ending point for both the upper and lower envelope models.
    u_x.append(len(s)-1)
    u_y.append(s[-1])

    l_x.append(len(s)-1)
    l_y.append(s[-1])
    #Fit suitable models to the data. Here I am using cubic splines, similarly to the MATLAB example given in the question.
    u_p = interp1d(u_x,u_y, kind = 'cubic',bounds_error = False, fill_value=0.0)
    l_p = interp1d(l_x,l_y,kind = 'cubic',bounds_error = False, fill_value=0.0)
    #Evaluate each model over the domain of (s)
    for k in range(0,len(s)):
        q_u[k] = u_p(k)
        q_l[k] = l_p(k)
    #Plot envelope:
    axs[tile].plot(q_u,'k-')   #upper envelope
    #axs[tile].plot(q_l,'g')   #lower envelope

    #plot sum of all time segments:
    #y = y1.add(y2)
    #y = y.add(y3)
    #axs[tile].plot(x,y, linewidth=0.75*lw, linestyle=linestyles[3], color="k")

    ##plot errorbars / error-regions:
    #err1 = df[df["wagon_type"] == wgType]["moving_time_std"].divide(3600)
    #err2 = df[df["wagon_type"] == wgType]["standing_time_std"].divide(3600)
    #err3 = df[df["wagon_type"] == wgType]["parking_time_std"].divide(3600)
    #axs[tile].fill_between(x, y1.subtract(err1), y1.add(err1), alpha=0.075, facecolor='#1f77b4')
    #axs[tile].fill_between(x, y2.subtract(err2), y2.add(err2), alpha=0.075, facecolor='#ff7f0e')
    #axs[tile].fill_between(x, y3.subtract(err3), y3.add(err3), alpha=0.075, facecolor='#2ca02c')


    axs[tile].set_ylim([0, 20])
    axs[tile].set_title("wagon Type "+str(tile+1), fontsize=18)
    #fig.legend(["mean moving time","mean standing time","mean parking time", "sum of all time segments"], loc='upper center', ncol=4)
    axs[tile].legend(["mean moving time","mean standing time","mean parking time", "envelope of moving time"], loc='upper center', ncol=4, prop={"size":16})

#some plot Settings:
plt.setp(axs, ylabel="time in hours (per day)", xlabel="day")
plt.xscale('linear')
plt.show()

## Histogram / Bar Plot: durchschnittliche tägliche Moving time eines jeden Wagon Typus

## Statistical description:
wgTypes = [i for i in range(0,9)]

movingTimeStats = []
standingTimeStats = []
parkingTimeStats = []

allData = []

for i, wgType in enumerate(wgTypes):
    if wgType == 0:
        df_temp = df
    else:
        df_temp = df[df.wagon_type == wgType]

    movingTimeStats.append([df_temp.moving_time_mean.mean(), df_temp.moving_time_median.median(), df_temp.moving_time_std.std()])
    standingTimeStats.append([df_temp.standing_time_mean.mean(), df_temp.standing_time_median.median(), df_temp.standing_time_std.std()])
    parkingTimeStats.append([df_temp.parking_time_mean.mean(), df_temp.parking_time_median.median(), df_temp.parking_time_std.std()])

    statDf = {"moving_time": movingTimeStats[i], "standing_time": standingTimeStats[i], "parking_time": parkingTimeStats[i]}

    statDf = pd.DataFrame(statDf, index=["mean", "median", "std"])
    statDf = statDf.divide(3600)
    #print("Wagon Type %i" %(wgType))
    #print(statDf)
    #print("")
    
    allData.append(statDf)

#allData

## Plot statistical description:
x = ["alle \n Wagons", "Typ 1", "Typ 2","Typ 3","Typ 4","Typ 5","Typ 6","Typ 7","Typ 8"]
y = []

for df_temp in allData:
    y.append(df_temp.moving_time[0])

x_pos = [1.6,3,4,5,6,7,8,9,10]
x_pos = [i*4 for i in x_pos]

fig, ax = plt.subplots(facecolor='w', edgecolor='k')

ax.bar(x_pos,y, width = 2.75)
ax.set_xticks(x_pos, x)
ax.axhline(y=y[0], color='r', linestyle='--')

ax.set_title("mittlere Bewegungszeit pro Tag")
ax.set_ylabel("Zeit in Stunden")
ax.grid(False)
plt.show()


fig.savefig(os.path.join("..","output","plots","group_01","mean_daily_moving_time_Bar_plot.png"), \
    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')
fig.savefig(os.path.join("..","output","plots","group_01","mean_daily_moving_time_Bar_plot.pdf"), \
    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')

# Plot: Wahrscheinlichkeitsdichteverteilung der durchschnittlichen täglichen moving time

# moving time per day distribution:

import scipy
import reliability      # pip install reliability

## Plot Histogramm / PDF
df = pd.read_csv(os.path.join("..", "output", "group_01", "all_wgID_dailyMean_moventTime_mvState=2.csv"))
x = df.mean_movingTimePerDay.divide(3600)

fig, ax = plt.subplots(facecolor='w', edgecolor='k')

n, bins, patches = ax.hist(x, bins=200, density=True, stacked=True, alpha=0.8, label="Daten")


# Weibull distribution
# https://reliability.readthedocs.io/en/latest/Equations%20of%20supported%20distributions.html
# 2 Parameter Weibull: alpha: scale parameter, beta: shape parameter, PDF = b/a*(x/a)**(b-1)*exp(-(x/a)**b), CDF = 1-exp(-(x/a)**b)
wbf = reliability.Fitters.Fit_Weibull_2P(failures=x.to_numpy(), show_probability_plot=False, print_results=False, method="MLE")  # fit the Weibull_3P distribution
print('Fit_Weibull_2P parameters:\nAlpha:', wbf.alpha, '\nBeta:', wbf.beta)
wbf.distribution.PDF(label='Weibull-Verteilung', linestyle='-', lw=3)  # plots to PDF of the fitted Weibull_2P

alpha = r"$\alpha$"
beta = r"$\beta$"
weibull_info = f"Weibull: \n {alpha}={np.round(wbf.alpha,2)} \n {beta}={np.round(wbf.beta,2)}"
plt.annotate(weibull_info, xy=(0.675, 0.575), xycoords='axes fraction', fontsize=15)

ax.set_xlim([0, 10])
ax.set_xlabel("tägliche gemittelte Bewegungszeit in Stunden")
ax.set_ylabel("Verteilung der Wagons")
ax.set_title("für alle Wagontypen (1-8)")

plt.grid(False)
plt.legend()
plt.show()

fig.savefig(os.path.join("..","output","plots","group_01","distribution_all_Wg_dailyMeanMovingTime.png"), \
    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')
fig.savefig(os.path.join("..","output","plots","group_01","distribution_all_Wg_dailyMeanMovingTime.pdf"), \
    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')

fz = 2*9
fz_lab = 2*8
fz_tick= 2*8
fz_leg = 2*6


# plot for/by each wagon type: total moving time distribution:
fig, axs = plt.subplots(2, 4, figsize=(14, 7), facecolor='w', edgecolor='k', constrained_layout=True, sharex=True, sharey=True) #make subplots
fig.subplots_adjust(hspace = .2, wspace=.15) #height and width space between the subfigures
axs = axs.ravel()


wgTypes = [i for i in range(1,9)]
for tile in range(8):
    wgType = wgTypes[tile]

    ## Plot Histogramm / PDF
    df = pd.read_csv(os.path.join("..", "output", "group_01", "all_wgID_dailyMean_moventTime_mvState=2.csv"))
    df = df[ df["wagon_type"] == wgType]
    x = df.mean_movingTimePerDay.divide(3600).to_numpy()
    #x = x*10

    n, bins, patches = axs[tile].hist(x, bins=200, density=True, stacked=True, alpha=0.8, label="Daten")

    wbf = reliability.Fitters.Fit_Weibull_2P(failures=x, show_probability_plot=False, print_results=False, method="MLE")  # fit the Weibull_2P distribution
    a, b = wbf.alpha, wbf.beta
    xx = np.linspace(0,100, 2000)
    axs[tile].plot( xx, b/a*(xx/a)**(b-1)*np.exp(-(xx/a)**b), label='Weibull-Funktion', linestyle='-', lw=3)

    alpha = r"$\alpha$"
    beta = r"$\beta$"
    weibull_info = f"Weibull: \n {alpha}={np.round(wbf.alpha,2)} \n {beta}={np.round(wbf.beta,2)}"
    if tile == 0:
        axs[tile].annotate(weibull_info, xy=(0.6, 0.4), xycoords='axes fraction', fontsize=fz)
    else:
        axs[tile].annotate(weibull_info, xy=(0.55, 0.6), xycoords='axes fraction', fontsize=fz)

    axs[tile].tick_params(axis='both', which='major', labelsize=fz_tick)
    axs[tile].tick_params(axis='both', which='minor', labelsize=fz_tick*0.8)
    axs[tile].set_xticks(np.arange(0, 11, step=2))

    axs[tile].set_title("Wagontyp "+str(tile+1), fontsize=fz)
    
    if tile==0:
        axs[tile].legend(fontsize=fz_leg)

    axs[tile].grid(False)
    axs[tile].set_xlim([0, 10])
    axs[tile].set_ylim([0, 1])

fig.add_subplot(111, frameon=False)
plt.tick_params(labelcolor='none', which='both', top=False, bottom=False, left=False, right=False)
plt.xlabel("tägliche gemittelte Bewegungszeit in Stunden", fontsize=fz_lab, labelpad=16)
plt.ylabel("Verteilung der Wagons", fontsize=fz_lab, labelpad=16)

#some plot Settings:
plt.xscale('linear')
plt.show()


fig.savefig(os.path.join("..","output","plots","group_01","distribution_each_WgType_dailyMeanMovingTime.png"), \
    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')
fig.savefig(os.path.join("..","output","plots","group_01","distribution_each_WgType_dailyMeanMovingTime.pdf"), \
    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')

## plot for/by each wagon type: total moving time distribution:
##fig, axs = plt.subplots(2, 4, figsize=(7, 5), facecolor='w', edgecolor='k', sharex = True, sharey = True, constrained_layout=True) #make subplots
#
#fig.subplots_adjust(hspace = .25, wspace=.15) #height and width space between the subfigures
#axs = axs.ravel()
#
#
#wgTypes = [i for i in range(1,9)]
#for tile in range(8):
#    wgType = wgTypes[tile]
#
#    ## Plot Histogramm / PDF
#    df = pd.read_csv(os.path.join("..", "output", "group_01", "all_wgID_dailyMean_moventTime_mvState=2.csv"))
#    df = df[ df["wagon_type"] == wgType]
#    x = df.mean_movingTimePerDay.divide(3600).to_numpy()
#    #x = x*10
#
#    n, bins, patches = axs[tile].hist(x, bins=200, density=True, stacked=True, alpha=0.8, label="Daten")
#
#    wbf = reliability.Fitters.Fit_Weibull_2P(failures=x, show_probability_plot=False, print_results=False, method="MLE")  # fit the Weibull_2P distribution
#    a, b = wbf.alpha, wbf.beta
#    xx = np.linspace(0,100, 2000)
#    axs[tile].plot( xx, b/a*(xx/a)**(b-1)*np.exp(-(xx/a)**b), label='Weibull-Verteilung', linestyle='-', lw=3)
#
#    alpha = r"$\alpha$"
#    beta = r"$\beta$"
#    weibull_info = f"Weibull: \n {alpha}={np.round(wbf.alpha,2)} \n {beta}={np.round(wbf.beta,2)}"
#    axs[tile].annotate(weibull_info, xy=(0.7, 0.625), xycoords='axes fraction', fontsize=15)
#
#    axs[tile].set_title("Wagon-Typ "+str(tile+1))
#    axs[tile].legend()
#    axs[tile].grid(False)
#    axs[tile].set_xlim([0, 10])
#    #axs[tile].set_ylim([0, 1])
#
##some plot Settings:
#plt.setp(axs, ylabel="Verteilung der Wagons", xlabel="tägliche gemittelte Bewegungszeit in Stunden")
#plt.xscale('linear')
#plt.show()
#
#fig.savefig(os.path.join("..","output","plots","group_01","distribution_each_WgType_dailyMeanMovingTime.png"), \
#    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')
#fig.savefig(os.path.join("..","output","plots","group_01","distribution_each_WgType_dailyMeanMovingTime.pdf"), \
#    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')

# Plot: Wahrscheinlichkeitsdichteverteilung des Anteils an der maximal gesamten Moving Time

# total moving time distribution:
import scipy
import reliability      # pip install reliability
from sklearn.preprocessing import MinMaxScaler

## Plot Histogramm / PDF
df = pd.read_csv(os.path.join("..", "output", "group_01", "all_wgID_total_moventTime_mvState=2.csv"))
x = df.total_movingTime.divide(df.total_movingTime.max()).to_numpy()
x = x*100

fig, ax = plt.subplots(facecolor='w', edgecolor='k')

n, bins, patches = ax.hist(x, bins=100, density=True, stacked=True, alpha=0.8, label="Daten")
print (np.sum(n*np.diff(bins)))

# Weibull distribution
# https://reliability.readthedocs.io/en/latest/Equations%20of%20supported%20distributions.html
# 2 Parameter Weibull: alpha: scale parameter, beta: shape parameter, PDF = b/a*(x/a)**(b-1)*exp(-(x/a)**b), CDF = 1-exp(-(x/a)**b)
wbf = reliability.Fitters.Fit_Weibull_2P(failures=x, show_probability_plot=False, print_results=False, method="MLE")  # fit the Weibull_2P distribution
print('Fit_Weibull_2P parameters: \
    \nAlpha:', wbf.alpha, "standard error:", wbf.alpha_SE, \
    '\nBeta:', wbf.beta, "standard error:", wbf.beta_SE)
a, b = wbf.alpha, wbf.beta
xx = np.linspace(0,100, 500)
plt.plot( xx, b/a*(xx/a)**(b-1)*np.exp(-(xx/a)**b), label='Weibull-Verteilung', linestyle='-', lw=3)

lognormfit = reliability.Fitters.Fit_Lognormal_2P(failures=x, show_probability_plot=False, print_results=False, method="MLE")

normfit = reliability.Fitters.Fit_Normal_2P(failures=x, show_probability_plot=False, print_results=False, method="MLE")

alpha = r"$\alpha$"
beta = r"$\beta$"
weibull_info = f"Weibull: \n {alpha}={np.round(wbf.alpha,2)} \n {beta}={np.round(wbf.beta,2)}"
plt.annotate(weibull_info, xy=(0.675, 0.575), xycoords='axes fraction', fontsize=15)

ax.set_xlabel("totale Bewegungszeit in \%")
ax.set_ylabel("Verteilung der Wagons")
ax.set_title("alle Wagontypen (1-8)")

plt.grid(False)
plt.legend()
plt.show()

fig.savefig(os.path.join("..","output","plots","group_01","distribution_all_Wg_totalMovingTime.png"), \
    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')
fig.savefig(os.path.join("..","output","plots","group_01","distribution_all_Wg_totalMovingTime.pdf"), \
    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')

fz = 2*9
fz_lab = 2*8
fz_tick= 2*8
fz_leg = 2*6


# plot for/by each wagon type: total moving time distribution:
fig, axs = plt.subplots(2, 4, figsize=(14, 7), facecolor='w', edgecolor='k', constrained_layout=True, sharex=True, sharey=True) #make subplots
fig.subplots_adjust(hspace = .2, wspace=.15) #height and width space between the subfigures
axs = axs.ravel()


wgTypes = [i for i in range(1,9)]
for tile in range(8):
    wgType = wgTypes[tile]

    ## Plot Histogramm / PDF
    df = pd.read_csv(os.path.join("..", "output", "group_01", "all_wgID_total_moventTime_mvState=2.csv"))
    max_total_movingTime = df.total_movingTime.max()
    df = df[ df["wagon_type"] == wgType]
    x = df.total_movingTime.divide(max_total_movingTime).to_numpy()
    x = x*100

    n, bins, patches = axs[tile].hist(x, bins=200, density=True, stacked=True, alpha=0.8, label="Daten")

    wbf = reliability.Fitters.Fit_Weibull_2P(failures=x, show_probability_plot=False, print_results=False, method="MLE")  # fit the Weibull_2P distribution
    a, b = wbf.alpha, wbf.beta
    xx = np.linspace(0,100, 2000)
    axs[tile].plot( xx, b/a*(xx/a)**(b-1)*np.exp(-(xx/a)**b), label='Weibull-Funktion', linestyle='-', lw=3)

    alpha = r"$\alpha$"
    beta = r"$\beta$"
    weibull_info = f"Weibull: \n {alpha}={np.round(wbf.alpha,2)} \n {beta}={np.round(wbf.beta,2)}"
    if tile == 0:
        axs[tile].annotate(weibull_info, xy=(0.6, 0.4), xycoords='axes fraction', fontsize=fz)
    else:
        axs[tile].annotate(weibull_info, xy=(0.55, 0.6), xycoords='axes fraction', fontsize=fz)

    axs[tile].tick_params(axis='both', which='major', labelsize=fz_tick)
    axs[tile].tick_params(axis='both', which='minor', labelsize=fz_tick*0.8)
    axs[tile].set_xticks(np.arange(0, 101, step=20))

    axs[tile].set_title("Wagontyp "+str(tile+1), fontsize=fz)
    
    if tile==0:
        axs[tile].legend(fontsize=fz_leg)

    axs[tile].grid(False)
    axs[tile].set_xlim([0, 100])
    axs[tile].set_ylim([0, 0.215])

fig.add_subplot(111, frameon=False)
plt.tick_params(labelcolor='none', which='both', top=False, bottom=False, left=False, right=False)
plt.xlabel("totale Bewegungszeit in \%", fontsize=fz_lab, labelpad=16)
plt.ylabel("Verteilung der Wagons", fontsize=fz_lab, labelpad=20)

#some plot Settings:
plt.xscale('linear')
plt.show()


fig.savefig(os.path.join("..","output","plots","group_01","distribution_each_wgType_totalMovingTime.png"), \
    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')
fig.savefig(os.path.join("..","output","plots","group_01","distribution_each_wgType_totalMovingTime.pdf"), \
    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')

## plot for/by each wagon type: total moving time distribution:
#fig, axs = plt.subplots(2, 4, figsize=(30, 10), facecolor='w', edgecolor='k') #make subplots
#fig.subplots_adjust(hspace = .25, wspace=.15) #height and width space between the subfigures
#axs = axs.ravel()
#
#
#wgTypes = [i for i in range(1,9)]
#for tile in range(8):
#    wgType = wgTypes[tile]
#
#    ## Plot Histogramm / PDF
#    df = pd.read_csv(os.path.join("..", "output", "group_01", "all_wgID_total_moventTime_mvState=2.csv"))
#    max_total_movingTime = df.total_movingTime.max()
#    df = df[ df["wagon_type"] == wgType]
#    x = df.total_movingTime.divide(max_total_movingTime).to_numpy()
#    x = x*100
#
#    n, bins, patches = axs[tile].hist(x, bins=100, density=True, stacked=True, alpha=0.8, label="Daten")
#
#    wbf = reliability.Fitters.Fit_Weibull_2P(failures=x, show_probability_plot=False, print_results=False, method="MLE")  # fit the Weibull_2P distribution
#    a, b = wbf.alpha, wbf.beta
#    xx = np.linspace(0,100, 2000)
#    axs[tile].plot( xx, b/a*(xx/a)**(b-1)*np.exp(-(xx/a)**b), label='Weibull-Verteilung', linestyle='-', lw=3)
#
#    alpha = r"$\alpha$"
#    beta = r"$\beta$"
#    weibull_info = f"Weibull: \n {alpha}={np.round(wbf.alpha,2)} \n {beta}={np.round(wbf.beta,2)}"
#    axs[tile].annotate(weibull_info, xy=(0.7, 0.625), xycoords='axes fraction', fontsize=15)
#
#    axs[tile].set_title("Wagon-Typ "+str(tile+1))
#    axs[tile].legend()
#    axs[tile].grid(False)
#
##some plot Settings:
#plt.setp(axs, ylabel="Verteilung der Wagons", xlabel="totale Bewegungszeit in \%")
#
#plt.xscale('linear')
#plt.show()
#
#fig.savefig(os.path.join("..","output","plots","group_01","distribution_each_wgType_totalMovingTime.png"), \
#    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')
#fig.savefig(os.path.join("..","output","plots","group_01","distribution_each_wgType_totalMovingTime.pdf"), \
#    bbox_inches='tight', dpi = 300,facecolor=fig.get_facecolor(), edgecolor='none')
