==> part_i_data_exploration_and_statistics/scripts_final/10_plot_distance_countries.ipynb <==
import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib as mplt
from matplotlib.patches import Patch

# Read data
filenamePath = os.path.join("..", "output", "group_02", "distance_countries.csv")
df = pd.read_csv(filenamePath)
df = df.dropna(inplace=False)
df

plt.style.use(['mystyle'])

# Plot Pie Chart with total distance traveled loaded/ unloaded
plt.style.use(['mystyle'])
pie_Chart = np.array([df["distance_empty"].sum(),df["distance_loaded"].sum()])

labels = ["Distance empty", "Distance loaded"]

fig, ax = plt.subplots(figsize=(3.5,3.5))
ax.pie(pie_Chart, labels = labels, autopct='%1.1f%%')
plt.show()
#plot_path = os.path.join("..", "output", "plots", "distance_traveled_piechart.pdf")
#fig.savefig(plot_path)

# Sort countries by total distance and take the highest 19, then merge the others into "Rest"
df_sorted = df.sort_values("distance_total", axis=0, ascending=False, na_position='last')
relevant_countries = df_sorted[0:15]

distance_total_rest = sum(df_sorted["distance_total"][16:])
distance_empty_rest = sum(df_sorted["distance_empty"][16:])
distance_loaded_rest = sum(df_sorted["distance_loaded"][16:])

df_temp = pd.DataFrame([[38, "Rest", distance_empty_rest, distance_loaded_rest, distance_total_rest]], 
                       columns=["Unnamed: 0", "country",  "distance_empty", "distance_loaded", "distance_total"])
relevant_countries = pd.concat([relevant_countries, df_temp])
relevant_countries

# barplot
labels = np.array(relevant_countries["country"])
width = 0.35       # the width of the bars

loaded = np.array(relevant_countries["distance_loaded"])
empty = np.array(relevant_countries["distance_empty"])

fig, ax = plt.subplots(figsize=(17.78,6))

ax.bar(labels, loaded, width, label='Beladen')
ax.bar(labels, empty , width, bottom=loaded,
       label='Unbeladen')

ax.set_ylabel('Strecke in 100 Mio. Kilometern')
ax.set_title('Insgesamt zurückgelegte Strecke in verschiedenen Ländern')
ax.legend()
fig.savefig(os.path.join("..","output","plots","group_02","distance_traveled_in_different_countries.png"), \
    bbox_inches='tight', dpi = 300,facecolor= 'w', edgecolor='none')
fig.savefig(os.path.join("..","output","plots","group_02","distance_traveled_in_different_countries.pdf"), \
    bbox_inches='tight', dpi = 300,facecolor= 'w', edgecolor='none')

# barplot foreign countries only
labels = np.array(relevant_countries["country"][1:])
width = 0.35       # the width of the bars

loaded = np.array(relevant_countries["distance_loaded"][1:])
empty = np.array(relevant_countries["distance_empty"][1:])

fig, ax = plt.subplots(figsize=(17.78,6))

ax.bar(labels, loaded, width, label='Beladen')
ax.bar(labels, empty , width, bottom=loaded,
       label='Unbeladen')

ax.set_ylabel('Strecke in 10 Mio. Kilometern')
ax.set_title('Insgesamt zurückgelegte Strecke im Ausland')
ax.legend()

plot_path = os.path.join("..", "output", "plots", "distance_traveled_in_foreign_countries.pdf")
fig.savefig(plot_path)

fig, ax = plt.subplots(figsize=(6,6))

size = 0.3
#sort distances, take 4 highest distances and sum up the rest
df = df.sort_values("distance_total", axis=0, ascending=False, na_position='last')

total_vals = np.array(df["distance_total"][0:7])
total_vals = np.append(total_vals,sum(df["distance_total"][7:]))

vals_empty = np.array(df["distance_empty"][0:7])
vals_empty = np.append(vals_empty,sum(df["distance_empty"][7:]))

vals_loaded = np.array(df["distance_loaded"][0:7])
vals_loaded = np.append(vals_loaded,sum(df["distance_loaded"][7:]))

vals = np.array([vals_loaded, vals_empty])

labels = np.array(['Deutschland', 'Italien', 'Niederlande', 'Österreich', 'Polen', 'Schweiz', 'Tschechien'])
labels = np.append(labels, "andere")

#Colormap
#https://matplotlib.org/stable/tutorials/colors/colormaps.html <- um Farben zu erändern
cmap_tab20c = plt.cm.tab20c
cmap_tab20b = plt.cm.tab20b
outer_colors = np.concatenate([cmap_tab20c([1,5,9,13]), cmap_tab20b([1,5,9,13])])
inner_colors = np.concatenate([cmap_tab20c([1,3,5,7,9,11,13,15]), cmap_tab20b([1,3,5,7,9,11,13,15])])

#Plot
ax.pie(total_vals, radius=1, colors=outer_colors, labels=labels,autopct='%1.1f%%', pctdistance=0.85,
       wedgeprops=dict(width=size, edgecolor='w'))

ax.pie(vals.flatten('F'), radius=1-size, colors=inner_colors,
       wedgeprops=dict(width=0.3, edgecolor='w'))

ax.set(aspect="equal", title='Zurückgelegte Distanzen in Ländern')

#Create Legend
legend_elements = [Patch(facecolor='grey', edgecolor='w',label='Beladen'),
                   Patch(facecolor='lightgrey', edgecolor='w',label='Leer')]

ax.legend(handles=legend_elements, loc='best', title="Innerer Kreis")

fig.savefig(os.path.join("..","output","plots","group_02","distance_in_countries_pie.png"), \
    bbox_inches='tight', dpi = 300,facecolor= 'w', edgecolor='none')
fig.savefig(os.path.join("..","output","plots","group_02","distance_in_countries_pie.pdf"), \
    bbox_inches='tight', dpi = 300,facecolor= 'w', edgecolor='none')

fig, ax = plt.subplots(figsize=(1,1))

size = 0.3

#sort distances, take 4 highest distances and sum up the rest
df = df.sort_values("distance_total", axis=0, ascending=False, na_position='last')

total_vals = np.array(df["distance_total"][0:7])
total_vals = np.append(total_vals,sum(df["distance_total"][7:]))

vals_empty = np.array(df["distance_empty"][0:7])
vals_empty = np.append(vals_empty,sum(df["distance_empty"][7:]))

vals_loaded = np.array(df["distance_loaded"][0:7])
vals_loaded = np.append(vals_loaded,sum(df["distance_loaded"][7:]))

vals = np.array([vals_loaded, vals_empty])

labels = np.char.array(['', '', '', '', '', '', ''])
labels = np.append(labels, "")

y = np.array(vals_loaded/total_vals)

porcent = 100.*y
labels = ['{0} -----  {1:1.2f} \%'.format(i,j) for i,j in zip(labels, porcent)]

#Colormap
#https://matplotlib.org/stable/tutorials/colors/colormaps.html <- um Farben zu erändern
cmap_tab20c = plt.cm.tab20c
cmap_tab20b = plt.cm.tab20b
outer_colors = np.concatenate([cmap_tab20c([1,5,9,13]), cmap_tab20b([1,5,9,13])])
inner_colors = np.concatenate([cmap_tab20c([1,3,5,7,9,11,13,15]), cmap_tab20b([1,3,5,7,9,11,13,15])])

#Plot

ax.set(aspect="equal", title='Zurückgelegte Distanzen der Wagontypen')

#Create Legend
patches, texts = plt.pie(y, colors=outer_colors, startangle=90, radius=1.2)
sort_legend = True
if sort_legend:
    patches, labels, dummy =  zip(*sorted(zip(patches, labels, y),
                                          key=lambda x: x[2],
                                          reverse=True))

plt.legend(patches, labels, loc=1, bbox_to_anchor=(-0.1, 1.),
           fontsize=10, title="l")

plt.show()

fig.savefig(os.path.join("..","output","plots","group_02","distance_in_countries_pie_legend_2.png"), \
    bbox_inches='tight', dpi = 300,facecolor= 'w', edgecolor='none')
#fig.savefig(os.path.join("..","output","plots","group_02","distance_in_countries_pie_legend.pdf"), \
#    bbox_inches='tight', dpi = 300,facecolor= 'w', edgecolor='none')


fig, ax = plt.subplots(figsize=(6,6))

size = 0.3
#sort distances, take 4 highest distances and sum up the rest
df = df.sort_values("distance_total", axis=0, ascending=False, na_position='last')

total_vals = np.array(df["distance_total"][1:8])
total_vals = np.append(total_vals,sum(df["distance_total"][8:]))

vals_empty = np.array(df["distance_empty"][1:8])
vals_empty = np.append(vals_empty,sum(df["distance_empty"][8:]))

vals_loaded = np.array(df["distance_loaded"][1:8])
vals_loaded = np.append(vals_loaded,sum(df["distance_loaded"][8:]))

vals = np.array([vals_loaded, vals_empty])

labels = np.array(['Italien', 'Niederlande', 'Österreich', 'Polen', 'Schweiz', 'Tschechien','Schweden'])
labels = np.append(labels, "andere")

#Colormap
#https://matplotlib.org/stable/tutorials/colors/colormaps.html <- um Farben zu erändern
cmap_tab20c = plt.cm.tab20c
cmap_tab20b = plt.cm.tab20b
outer_colors = np.concatenate([cmap_tab20c([5,9,13,1]), cmap_tab20b([5,9,13,1])])
inner_colors = np.concatenate([cmap_tab20c([5,7,9,11,13,15,1,3]), cmap_tab20b([5,7,9,11,13,15,1,3])])

#Plot
ax.pie(total_vals, radius=1, colors=outer_colors, labels=labels, autopct='%1.1f%%', pctdistance=0.85,
       wedgeprops=dict(width=size, edgecolor='w'))

ax.pie(vals.flatten('F'), radius=1-size, colors=inner_colors,
       wedgeprops=dict(width=0.3, edgecolor='w'))

ax.set(aspect="equal", title='Zurückgelegte Distanzen in Ländern außer Deutschland')

#Create Legend
legend_elements = [Patch(facecolor='grey', edgecolor='w',label='Beladen'),
                   Patch(facecolor='lightgrey', edgecolor='w',label='Leer')]

ax.legend(handles=legend_elements, loc='best', title="Innerer Kreis")

fig.savefig(os.path.join("..","output","plots","group_02","distance_in_countries_pie_not_D.png"), \
    bbox_inches='tight', dpi = 300,facecolor= 'w', edgecolor='none')
fig.savefig(os.path.join("..","output","plots","group_02","distance_in_countries_pie_not_D.pdf"), \
    bbox_inches='tight', dpi = 300,facecolor= 'w', edgecolor='none')
