==> part_i_data_exploration_and_statistics/scripts_final/11_plot_distance_by_type.ipynb <==
import pandas as pd
import numpy as np

import os
import dask.dataframe as dd
import matplotlib as mpl
import matplotlib.pyplot as plt
import seaborn as sns
from matplotlib.patches import Patch

from os.path import join

### Einlesen der Daten
filenamePath = os.path.join("..", "output", "group_02", "distance_all.csv")
df = pd.read_csv(filenamePath)
df

plt.style.use(['mystyle'])

#Plot Pie Chart Mean distance loaded/ unloaded

pie_Chart = np.array([df["distance_empty"].sum(),df["distance_loaded"].sum()])

#Colormap
#https://matplotlib.org/stable/tutorials/colors/colormaps.html <- um Farben zu erändern
cmap_tab20c = plt.cm.tab20c
cmap_tab20b = plt.cm.tab20b
colors = cmap_tab20c([1,5])

labels = ["Summe Distanz Leer", "Summe Distanz Beladen"]

plt.pie(pie_Chart, colors=colors, labels = labels, autopct='%1.1f%%')
plt.savefig(os.path.join("..","output","plots","group_02","distance_pie_chart.png"), facecolor = "w", dpi = 300)
plt.savefig(os.path.join("..","output","plots","group_02","distance_pie_chart.pdf"), facecolor = "w", dpi = 300)
plt.show()
#fig[0].savefig(os.path.join("..","output","plots","group_02","distance_pie_chart.pdf"), \
#    bbox_inches='tight', dpi = 300, edgecolor='none')
#type(fig[1])

## Plot statistical description:
x = ["Typ 1", "Typ 2","Typ 3","Typ 4","Typ 5","Typ 6","Typ 7","Typ 8"]

y = df['distance_total']/df['distance_total'].mean()*100

x_pos = [1,2,3,4,5,6,7,8]
x_pos = [i*4 for i in x_pos]

fig, ax = plt.subplots()

ax.bar(x_pos,y, width = 2.75)
ax.set_xticks(x_pos, x)

ax.set_title("Zurückgelegte Distanz im Vergleich pro Wagontyp im Vergleich zum Durchschnitt")
ax.set_ylabel("Distanz durch Durchschnittsdistanz in \%")
ax.grid(False)
ax.axhline(100, color='r', linestyle='--')

plt.show()


fig.savefig(os.path.join("..","output","plots","group_02","percentage_distance_travelled_Bar_plot.png"), \
    bbox_inches='tight', dpi = 300,facecolor= 'w', edgecolor='none')
fig.savefig(os.path.join("..","output","plots","group_02","percentage_distance_travelled_Bar_plot.pdf"), \
    bbox_inches='tight', dpi = 300,facecolor= 'w', edgecolor='none')


fig, ax = plt.subplots(figsize=(6,6))

size = 0.3

total_vals = np.array(df["distance_total"])

vals_empty = np.array(df['distance_empty'])
vals_loaded = np.array(df['distance_loaded'])
vals = np.array([vals_loaded, vals_empty])

labels = ["Typ 1", "Typ 2", "Typ 3", "Typ 4", "Typ 5", "Typ 6", "Typ 7", "Typ 8"]

#Colormap
#https://matplotlib.org/stable/tutorials/colors/colormaps.html <- um Farben zu erändern
cmap_tab20c = plt.cm.tab20c
cmap_tab20b = plt.cm.tab20b
outer_colors = np.concatenate([cmap_tab20c([5,9,13,1]), cmap_tab20b([5,9,13,1])])
inner_colors = np.concatenate([cmap_tab20c([5,7,9,11,13,15,1,3]), cmap_tab20b([5,7,9,11,13,15,1,3])])

#Plot
ax.pie(total_vals, radius=1, colors=outer_colors, labels=labels, autopct='%1.1f%%', pctdistance=0.85,
       wedgeprops=dict(width=size, edgecolor='w'))

ax.pie(vals.flatten('F'), radius=1-size, colors=inner_colors,
       wedgeprops=dict(width=0.3, edgecolor='w'))

ax.set(aspect="equal", title='Zurückgelegte Distanzen der Wagontypen')

#Create Legend
legend_elements = [Patch(facecolor='grey', edgecolor='w',label='Beladen'),
                   Patch(facecolor='lightgrey', edgecolor='w',label='Leer')]

ax.legend(handles=legend_elements, loc=1, title="Innerer Kreis")
plt.show()

#fig.savefig(os.path.join("..","output","plots","group_02","mean_wagontypen_distances.png"), \
#    bbox_inches='tight', dpi = 300, facecolor = "w", edgecolor='none')
#fig.savefig(os.path.join("..","output","plots","group_02","mean_wagontypen_distances.pdf"), \
#    bbox_inches='tight', dpi = 300, facecolor = "w", edgecolor='none')
