==> part_i_data_exploration_and_statistics/scripts_final/16_AnaMap.ipynb <==
### Zuordnung der Provider zu den Wagentypen

### Importieren der verwendeten Bibilotheken
import os
import glob
from collections import Counter

import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import matplotlib.patches as mpatches

plt.style.use(['mystyle'])

#wagon_typeList = np.sort(df["wagon_type"].drop_duplicates().to_numpy())
#wagon_typeList

pathToPickleMapping = os.path.join("..", "..", "data", "mappingdf.pickle")
mappingdf = pd.read_pickle(pathToPickleMapping)

mappingdf

providerList = np.sort(mappingdf["provider"].drop_duplicates().to_numpy())
providerList = np.delete(providerList ,0)


dfTemps = []
for provider in providerList:
    dfTemps.append(mappingdf[mappingdf['provider']==provider])


#plt.style.use('seaborn-deep')
fig, ax = plt.subplots(nrows=1, ncols=1, constrained_layout=True, figsize=(3.5,3))
bins=np.arange(10)-0.5
bins = np.delete(bins, 0)
ax.hist([dfTemps[0]['wagon_type'], dfTemps[1]['wagon_type'], dfTemps[2]['wagon_type']], stacked = True, label = ['Provider 32', 'Provider 35', 'Provider 37'], rwidth=0.7, bins=bins)
# ax.set_title('Anzahl Wagons, geordnet nach Wagontypen,\n eingeteilt nach Provider')
ax.set_ylabel('Anzahl Wagons')
ax.set_xlabel('Wagontyp')
ax.grid(False)
legend_box = (0.145, 1.02, 0.855, 0.7)
legend = fig.legend(fontsize='small', bbox_to_anchor=legend_box, ncol=3, loc=3, borderaxespad=0.0, mode='expand', handlelength=1)
